+++ VehicleFinder/Entities/Vehicle.cs
@@ -14,7 +14,7 @@ namespace VehicleFinder.Entities
         public int NumberOfPreviousOwners { get; set; }
         public ShifterType ShifterType { get; set; }
         public int GearCount { get; set; }
-        public decimal AverageConsumption { get; set; }
+        public double AverageConsumption { get; set; }
         public string EngineId { get; set; }
         public Engine? Engine { get; set; }
         public string BodyId { get; set; }
